rpng: avoid signed overflow in rpng_dword_be
buf[0] promotes to int, so a leading byte >= 0x80 shifted left by 24
overflows a 32-bit signed int.  UBSan reports it on every chunk header
whose first length byte has the high bit set:

  rpng.c:296:19: runtime error: left shift of 255 by 24 places cannot
  be represented in type 'int'

Visible in the CI sample run via rpng_chunk_overflow_test, which feeds
chunk_size=0xFFFFFFF8 deliberately.  Widen each byte to uint32_t before
shifting; the result is unchanged on every two's-complement target.

+++ libretro-common/formats/png/rpng.c
@@ -293,7 +293,12 @@ static const struct adam7_pass rpng_passes[] = {
 
 static INLINE uint32_t rpng_dword_be(const uint8_t *buf)
 {
-   return (buf[0] << 24) | (buf[1] << 16) | (buf[2] << 8) | (buf[3] << 0);
+   /* Each byte promotes to int, so a leading byte >= 0x80 shifted left by
+    * 24 overflows a 32-bit signed int.  That is undefined behavior, and
+    * UBSan flags it on the chunk-header path for any chunk whose length
+    * or type byte has the high bit set.  Widen to uint32_t first. */
+   return  ((uint32_t)buf[0] << 24) | ((uint32_t)buf[1] << 16)
+         | ((uint32_t)buf[2] <<  8) | ((uint32_t)buf[3] <<  0);
 }
 
 static INLINE uint16_t rpng_word_be(const uint8_t *buf)
